pyproject.toml
[tool.poetry]
name = "aquapod"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.4.0"}
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
